requirements.txt
# PPKS NLP Project — Python Dependencies
# Install dengan: pip install -r requirements.txt

# ── Data & Utilities ─────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1          # Untuk menyimpan API key dengan aman

# ── Twitter Scraping ─────────────────────────────────────────
tweepy==4.14.0                # Library resmi Twitter API

# ── NLP & Text Processing ────────────────────────────────────
PySastrawi==1.2.0             # Stemmer Bahasa Indonesia
nltk==3.8.1                   # Toolkit NLP dasar
emoji==2.12.1                 # Konversi emoji ke teks
scikit-learn==1.4.2           # TF-IDF, LDA, evaluasi model

# ── Deep Learning / IndoBERT ─────────────────────────────────
# torch versi CPU (lebih ringan, tidak butuh GPU)
torch==2.3.0
transformers==4.41.2          # HuggingFace untuk IndoBERT
sentencepiece==0.2.0          # Tokenizer IndoBERT

# ── Visualisasi ───────────────────────────────────────────────
matplotlib==3.9.0
seaborn==0.13.2
wordcloud==1.9.3

# ── Output & Reporting ───────────────────────────────────────
jinja2==3.1.4                 # Template HTML report